requirements.txt
Pillow==10.0.1
numpy==1.26.0
opencv-python~=4.8.1.78
pytesseract==0.3.10
win32gui